pyproject.toml
[project]
name = "mentorship-backend-learning-journey-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "pydantic-settings>=2.13.1",
    "sqlalchemy[asyncio]>=2.0.48",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
pythonpath = ["services/api", "."]
filterwarnings = [
    "ignore::DeprecationWarning",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
